Enhancement: Turn AbstractTestClassTestCase into assertion
