+++ src/TestHelper.php
@@ -10,6 +10,7 @@
 namespace Refinery29\Test\Util;
 
 use Faker\Factory;
+use Zend\File;
 
 trait TestHelper
 {
@@ -151,6 +152,103 @@ trait TestHelper
             'Failed to assert that "%s" implements "%s"',
             $className,
             $interfaceName
+        ));
+    }
+
+    /**
+     * @param string   $path
+     * @param string[] $excludeDirectories
+     *
+     * @throws \InvalidArgumentException
+     */
+    final protected function assertClassesAreAbstractOrFinal($path, array $excludeDirectories = [])
+    {
+        if (!\is_string($path)) {
+            throw new \InvalidArgumentException(\sprintf(
+                'Path needs to be specified as a string, got "%s".',
+                \is_object($path) ? \get_class($path) : \gettype($path)
+            ));
+        }
+
+        if (!\is_dir($path)) {
+            throw new \InvalidArgumentException(\sprintf(
+                'Path needs to be specified as an existing directory, got "%s".',
+                $path
+            ));
+        }
+
+        \array_walk($excludeDirectories, function ($excludeDirectory) use ($path) {
+            if (!\is_string($excludeDirectory)) {
+                throw new \InvalidArgumentException(\sprintf(
+                    'Exclude directory needs to be specified as a string, got "%s".',
+                    \is_object($excludeDirectory) ? \get_class($excludeDirectory) : \gettype($excludeDirectory)
+                ));
+            }
+
+            if (!\is_dir($path . DIRECTORY_SEPARATOR . $excludeDirectory)) {
+                throw new \InvalidArgumentException(\sprintf(
+                    'Exclude directory needs to point to an existing directory within "%s", got "%s".',
+                    $path,
+                    $excludeDirectory
+                ));
+            }
+        });
+
+        $classFileLocator = new File\ClassFileLocator($path);
+
+        $classFiles = \iterator_to_array(
+            $classFileLocator,
+            false
+        );
+
+        $excludePaths = \array_map(function ($excludeDirectory) use ($path) {
+            return \realpath($path . DIRECTORY_SEPARATOR . $excludeDirectory);
+        }, $excludeDirectories);
+
+        $classNames = \array_reduce(
+            $classFiles,
+            function (array $classNames, File\PhpClassFile $file) use ($excludePaths) {
+                $realPath = $file->getRealPath();
+                foreach ($excludePaths as $excludePath) {
+                    if (\strpos($realPath, $excludePath) === 0) {
+                        return $classNames;
+                    }
+                }
+
+                return array_merge(
+                    $classNames,
+                    $file->getClasses()
+                );
+            },
+            []
+        );
+
+        $this->assertGreaterThan(0, \count($classNames), \sprintf(
+            'Could not find any relevant PHP files in path "%s"%s.',
+            $path,
+            \count($excludeDirectories) === 0 ? '' : \sprintf(
+                ' excluding "%s"',
+                \implode('", "', $excludeDirectories)
+            )
+        ));
+
+        $classNamesNeitherAbstractNorFinal = \array_filter($classNames, function ($className) {
+            $reflection = new \ReflectionClass($className);
+
+            if ($reflection->isTrait()
+                || $reflection->isInterface()
+                || $reflection->isAbstract()
+                || $reflection->isFinal()
+            ) {
+                return false;
+            }
+
+            return true;
+        });
+
+        $this->assertEmpty($classNamesNeitherAbstractNorFinal, \sprintf(
+            'Failed to assert that the following classes are abstract or final: %s',
+            PHP_EOL . ' - ' . \implode(PHP_EOL . ' - ', $classNamesNeitherAbstractNorFinal)
         ));
     }
 }
